moving state functions to engine.ts.. only canvas/render/draw functions here
