+++ src/game/PhaserGame.svelte
@@ -1,46 +1,39 @@
 <script context="module" lang="ts">
+	import type { Game, Scene } from 'phaser';
 
-    import type { Game, Scene } from "phaser";
-
-    export type TPhaserRef = {
-        game: Game | null,
-        scene: Scene | null
-    };
-
+	export type TPhaserRef = {
+		game: Game | null;
+		scene: Scene | null;
+	};
 </script>
 
 <script lang="ts">
+	import { onMount } from 'svelte';
+	import StartGame from './main';
+	import { EventBus } from './EventBus';
 
-    import { onMount } from "svelte";
-    import StartGame from "./main";
-    import { EventBus } from './EventBus';
+	export let phaserRef: TPhaserRef = {
+		game: null,
+		scene: null
+	};
 
-    export let phaserRef: TPhaserRef = {
-        game: null,
-        scene: null
-    };
+	export let currentActiveScene: (scene: Scene) => void | undefined;
 
-    export let currentActiveScene: (scene: Scene) => void | undefined;
+	onMount(() => {
+		phaserRef.game = StartGame('game-container');
 
-    onMount(() => {
+		EventBus.on('current-scene-ready', (scene_instance: Scene) => {
+			phaserRef.scene = scene_instance;
 
-        phaserRef.game = StartGame("game-container");
-
-        EventBus.on('current-scene-ready', (scene_instance: Scene) => {
-
-            phaserRef.scene = scene_instance;
-
-            if(currentActiveScene)
-            {
-                
-                currentActiveScene(scene_instance);
-                
-            }
-
-        });
-
-    });
-
+			if (currentActiveScene) {
+				currentActiveScene(scene_instance);
+			}
+		});
+	});
 </script>
 
-<div id="game-container"></div>
+<div id="game-container">
+	<div id="worldCanvas">
+		<div id="currentCityCanvas"></div>
+	</div>
+</div>
