+++ src/lib/engine/engine.ts
@@ -1,184 +1,206 @@
 import { gameParams, gameState, loading } from '../../stores.ts';
-import {mean} from 'mathjs';
+import { mean } from 'mathjs';
 
 //TODO: move most of the game mechanics logic that's not UI here instead of context menu, canvas
 //svelte components
-let genMap = function(locationOptions, buildingPositionOpts) {
-      console.debug("genMap called with arguments");
-      console.debug(locationOptions);
-      console.debug(buildingPositionOpts);
-      //let city_cnts = Math.floor(Math.random() * 10);
-      let wmap = new Object();
-      let city_cnts = 1;
-      wmap.cities = new Array();
-      console.debug("Generating %s random cities ", city_cnts)
-      for(let i = 0; i < city_cnts; i++){
-        let cityObj = new Object();
-        cityObj.locations = new Array();
-        // generate random location/buildings per city and a bot for each location
-        let location_cnts = Math.floor(Math.random()*10);
+let genMap = function (locationOptions, buildingPositionOpts) {
+	console.debug('genMap called with arguments');
+	console.debug(locationOptions);
+	console.debug(buildingPositionOpts);
+	//let city_cnts = Math.floor(Math.random() * 10);
+	let wmap = new Object();
+	let city_cnts = 1;
+	wmap.cities = new Array();
+	console.debug('Generating %s random cities ', city_cnts);
+	for (let i = 0; i < city_cnts; i++) {
+		let cityObj = new Object();
+		cityObj.locations = new Array();
+		// generate random location/buildings per city and a bot for each location
+		let location_cnts = Math.floor(Math.random() * 10);
 
-        console.debug("Generating %s random buildings for city no: %s", location_cnts, i);
-        for(let j = 0; j < location_cnts; j++){
-          let randIdx = Math.floor(Math.random() * locationOptions.length);
-          let chosen_loc = locationOptions[randIdx];
-          chosen_loc.pos = buildingPositionOpts.splice(Math.floor(Math.random()* buildingPositionOpts.length), 1);
-          let bot = addBot();
-          cityObj.locations.push({loc: chosen_loc,
-                                  bots: [bot] });
-        }
-        wmap.cities.push(cityObj);
-      }
-      console.log("Finished generating Map");
-      console.debug(wmap);
-      return wmap;
+		console.debug('Generating %s random buildings for city no: %s', location_cnts, i);
+		for (let j = 0; j < location_cnts; j++) {
+			let randIdx = Math.floor(Math.random() * locationOptions.length);
+			let chosen_loc = locationOptions[randIdx];
+			chosen_loc.pos = buildingPositionOpts.splice(
+				Math.floor(Math.random() * buildingPositionOpts.length),
+				1
+			);
+			let bot = addBot();
+			cityObj.locations.push({ loc: chosen_loc, bots: [bot] });
+		}
+		wmap.cities.push(cityObj);
+	}
+	console.log('Finished generating Map');
+	console.debug(wmap);
+	return wmap;
 };
 
-let addBot= function (){
-      // Generate some random initial players for each location
-      const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-      let counter = 0;
-      let rand_id = '', rand_name = '';
-      length = Math.floor(Math.random() * 10);
-      while (counter < length) {
-        rand_id += characters.charAt(Math.floor(Math.random() * characters.Length));
-        rand_name += characters.charAt(Math.floor(Math.random() * characters.Length));
-        counter += 1;
-      }
-      let newUserObj = {
-          id: rand_id,
-          name: rand_name,
-          health: Math.floor(Math.random() * 100),
-          energy: {
-              social: Math.floor(Math.random() * 100),
-              focus: Math.floor(Math.random() * 100),
-              restless: Math.floor(Math.random() * 100),
-          },
-          neuro: {
-            interest: Math.floor(Math.random() * 100),
-            hyperfocused: Math.floor(Math.random() * 100),
-            asocial: Math.floor(Math.random() * 100),
-            mirror: Math.floor(Math.random() * 100),
-          },
-          gender: {
-            conform: Math.floor(Math.random() * 100),
-            weak: Math.floor(Math.random() * 100),
-            emo: Math.floor(Math.random() * 100),
-          },
-          sexuality: {
-            hetero: Math.floor(Math.random() * 100),
-            homo: Math.floor(Math.random() * 100),
-            ace: Math.floor(Math.random() * 100),
-            pan: Math.floor(Math.random() * 100),
-          },
-          leadership: {
-            leader: Math.floor(Math.random() * 100),
-            follower: Math.floor(Math.random() * 100),
-            ownway: Math.floor(Math.random() * 100),
-          },
-          social: {
-            listener: Math.floor(Math.random() * 100),
-            talker: Math.floor(Math.random() * 100),
-            observer: Math.floor(Math.random() * 100),
-          },
-         alertLevel: Math.floor(Math.random() * 100),
-      };
-      return newUserObj;
-      //gameState.no_of_users += 1;
-      //gameState.allUsers.push(newUserObj);
-      //gameState.locationUserMap[location].push(newUserObj.id);
-    };
+let addBot = function () {
+	// Generate some random initial players for each location
+	const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+	let counter = 0;
+	let rand_id = '',
+		rand_name = '';
+	length = Math.floor(Math.random() * 10);
+	while (counter < length) {
+		rand_id += characters.charAt(Math.floor(Math.random() * characters.Length));
+		rand_name += characters.charAt(Math.floor(Math.random() * characters.Length));
+		counter += 1;
+	}
+	let newUserObj = {
+		id: rand_id,
+		name: rand_name,
+		health: Math.floor(Math.random() * 100),
+		energy: {
+			social: Math.floor(Math.random() * 100),
+			focus: Math.floor(Math.random() * 100),
+			restless: Math.floor(Math.random() * 100)
+		},
+		neuro: {
+			interest: Math.floor(Math.random() * 100),
+			hyperfocused: Math.floor(Math.random() * 100),
+			asocial: Math.floor(Math.random() * 100),
+			mirror: Math.floor(Math.random() * 100)
+		},
+		gender: {
+			conform: Math.floor(Math.random() * 100),
+			weak: Math.floor(Math.random() * 100),
+			emo: Math.floor(Math.random() * 100)
+		},
+		sexuality: {
+			hetero: Math.floor(Math.random() * 100),
+			homo: Math.floor(Math.random() * 100),
+			ace: Math.floor(Math.random() * 100),
+			pan: Math.floor(Math.random() * 100)
+		},
+		leadership: {
+			leader: Math.floor(Math.random() * 100),
+			follower: Math.floor(Math.random() * 100),
+			ownway: Math.floor(Math.random() * 100)
+		},
+		social: {
+			listener: Math.floor(Math.random() * 100),
+			talker: Math.floor(Math.random() * 100),
+			observer: Math.floor(Math.random() * 100)
+		},
+		alertLevel: Math.floor(Math.random() * 100)
+	};
+	return newUserObj;
+	//gameState.no_of_users += 1;
+	//gameState.allUsers.push(newUserObj);
+	//gameState.locationUserMap[location].push(newUserObj.id);
+};
 
-let updateGameState = function(userStats, gameParams,
-                                currentLocation, spaceHoldingDrainer ) {
-        console.debug("update game state at");
-        console.log(currentLocation);
-        console.log(userStats.user.health);
-        console.log(userStats.user.energy);
-        console.log(userStats.user.alertLevel);
+let updateGameState = function (userStats, gameParams, currentLocation, spaceHoldingDrainer) {
+	console.debug('update game state at');
+	console.log(currentLocation);
+	console.log(userStats.user.health);
+	console.log(userStats.user.energy);
+	console.log(userStats.user.alertLevel);
 
-        let user_health = userStats.user.health;
-        let user_energy = userStats.user.energy;
-        let user_alertness = userStats.user.alertLevel;
-        // Calculate user parameters based on the locations
-        user_health += gameParams.locations[currentLocation].drain_rate.health * user_health;
-        user_alertness += gameParams.locations[currentLocation].drain_rate.alertness * user_alertness;
-        user_energy = {
-            social: user_energy.social + gameParams.locations[currentLocation].drain_rate.energy * user_energy.social,
-            focus: user_energy.focus + gameParams.locations[currentLocation].drain_rate.energy * user_energy.focus,
-            restless: user_energy.restless + gameParams.locations[currentLocation].drain_rate.energy * user_energy.restless,
-            }
-        if (userStats.locationUserMap[currentLocation].length > 0) {
-            // Calculate user parameters based on the nearby people
-            user_health += mean(userStats.locationUserMap[currentLocation].map( a=> a.health));
-            //TODO: explore the other attribute values and how they affect/impact each other's values in
-            //a group setting
-            user_energy = {
-                social: user_energy.social +
-                            // average social energy, probbaly wrong logic
-                            mean(userStats.locationUserMap[currentLocation].map( a=> a.energy.social)) +
-                            // Idea being holding space for too many people can drain one's social battery
-                            userStats.locationUserMap[currentLocation].length * spaceHoldingDrainer,
-                focus: user_energy.focus+
-                            // average focus energy, probbaly wrong logic
-                            mean(userStats.locationUserMap[currentLocation].map( a=> a.energy.focus)) + // Idea being holding space for too many people can drain one's social battery
-                            userStats.locationUserMap[currentLocation].length * spaceHoldingDrainer,
-                restless: user_energy.restless +
-                            // average restless energy, probbaly wrong logic
-                            mean(userStats.locationUserMap[currentLocation].map( a=> a.energy.restless)) +
-                            // Idea being holding space for too many people can drain one's social battery
-                            userStats.locationUserMap[currentLocation].length * spaceHoldingDrainer,
+	let user_health = userStats.user.health;
+	let user_energy = userStats.user.energy;
+	let user_alertness = userStats.user.alertLevel;
+	// Calculate user parameters based on the locations
+	user_health += gameParams.locations[currentLocation].drain_rate.health * user_health;
+	user_alertness += gameParams.locations[currentLocation].drain_rate.alertness * user_alertness;
+	user_energy = {
+		social:
+			user_energy.social +
+			gameParams.locations[currentLocation].drain_rate.energy * user_energy.social,
+		focus:
+			user_energy.focus +
+			gameParams.locations[currentLocation].drain_rate.energy * user_energy.focus,
+		restless:
+			user_energy.restless +
+			gameParams.locations[currentLocation].drain_rate.energy * user_energy.restless
+	};
+	if (userStats.locationUserMap[currentLocation].length > 0) {
+		// Calculate user parameters based on the nearby people
+		user_health += mean(userStats.locationUserMap[currentLocation].map((a) => a.health));
+		//TODO: explore the other attribute values and how they affect/impact each other's values in
+		//a group setting
+		user_energy = {
+			social:
+				user_energy.social +
+				// average social energy, probbaly wrong logic
+				mean(userStats.locationUserMap[currentLocation].map((a) => a.energy.social)) +
+				// Idea being holding space for too many people can drain one's social battery
+				userStats.locationUserMap[currentLocation].length * spaceHoldingDrainer,
+			focus:
+				user_energy.focus +
+				// average focus energy, probbaly wrong logic
+				mean(userStats.locationUserMap[currentLocation].map((a) => a.energy.focus)) + // Idea being holding space for too many people can drain one's social battery
+				userStats.locationUserMap[currentLocation].length * spaceHoldingDrainer,
+			restless:
+				user_energy.restless +
+				// average restless energy, probbaly wrong logic
+				mean(userStats.locationUserMap[currentLocation].map((a) => a.energy.restless)) +
+				// Idea being holding space for too many people can drain one's social battery
+				userStats.locationUserMap[currentLocation].length * spaceHoldingDrainer
+		};
 
-                }
+		user_alertness += userStats.locationUserMap[currentLocation].length * spaceHoldingDrainer;
+	}
 
-            user_alertness += userStats.locationUserMap[currentLocation].length * spaceHoldingDrainer;
-        }
+	return {
+		health: user_health,
+		energy: user_energy,
+		alertness: user_alertness,
+		locationUserMap: userStats.locationUserMap
+	};
+};
 
-        return {
-                health: user_health,
-                energy: user_energy,
-                alertness: user_alertness,
-                locationUserMap: userStats.locationUserMap,
-            };
-        };
+let genBots = function (locations) {
+	let crowd;
+	let allUsers = new Array();
+	let locationUserMap = new Object();
+	for (const loc in locations) {
+		let loc2 = loc;
+		switch (loc) {
+			case 'home':
+				crowd = 2;
+			case 'university':
+				crowd = 20;
+			case 'library':
+				crowd = 5;
+			case 'suicide_park':
+				crowd = 5;
+			case 'dance':
+				crowd = 10;
+			default:
+				crowd = 1;
+		}
+		console.log(loc2);
+		for (let i = 0; i < crowd; i++) {
+			let newBot;
+			newBot = addBot();
+			locationUserMap[loc2] = newBot;
+			allUsers.push(newBot.id);
+		}
+	}
+	return { allUsers: allUsers, locationUserMap: locationUserMap };
+};
 
-
-let genBots = function(locations) {
-    let crowd;
-    let allUsers = new Array();
-    let locationUserMap = new Object();
-    for (const loc in locations) {
-        let loc2 = loc;
-        switch(loc) {
-            case 'home':
-                crowd = 2;
-            case 'university':
-                crowd = 20;
-            case 'library':
-                crowd = 5;
-            case 'suicide_park':
-                crowd = 5;
-            case 'dance':
-                crowd = 10;
-            default:
-                crowd = 1;
-        }
-      console.log(loc2);
-      for (let i = 0; i < crowd; i++) {
-          let newBot;
-          newBot = addBot();
-          locationUserMap[loc2] = newBot;
-          allUsers.push(newBot.id);
-          }
-    }
-    return { allUsers: allUsers,
-             locationUserMap: locationUserMap,
-    }
+function initializeGameState(canvas) {
+	const style = canvas.style('.mycolor', { color: 'pink' });
+	//$spaceHoldingDrainer = socialDrainMultiplier($gameState.user);
+	canvas.add(style);
+	let genMap;
+	genMap = engine.gm(locationSpecificParams, buildingPositions);
+	for (let i = 0; i < genMap.cities.length; i++) {
+		let cityObj = genMap.cities[i];
+		//drawCity(canvas, cityObj);
+		gameState.worldmap.cities.push(cityObj);
+	}
+	loading.set(false);
 }
+
 const engine = {
-                 ab: addBot,
-                 ugs: updateGameState,
-                 gb: genBots,
-                 gm: genMap,
-                };
+	ab: addBot,
+	ugs: updateGameState,
+	gb: genBots,
+	gm: genMap
+};
 export { engine };
